+++ src/command_modules/azure-cli-role/azure/cli/command_modules/role/commands.py
@@ -88,7 +88,7 @@ def load_command_table(self, _):
         g.custom_command('create', 'create_application')
         g.custom_command('delete', 'delete_application')
         g.custom_command('list', 'list_apps')
-        g.custom_command('show', 'show_application')
+        g.custom_show_command('show', 'show_application')
         g.generic_update_command('update', setter_name='patch_application', setter_type=role_custom,
                                  getter_name='show_application', getter_type=role_custom,
                                  custom_func_name='update_application', custom_func_type=role_custom)
@@ -97,7 +97,7 @@ def load_command_table(self, _):
         g.custom_command('create', 'create_service_principal')
         g.custom_command('delete', 'delete_service_principal')
         g.custom_command('list', 'list_sps', client_factory=get_graph_client_service_principals)
-        g.custom_command('show', 'show_service_principal', client_factory=get_graph_client_service_principals)
+        g.custom_show_command('show', 'show_service_principal', client_factory=get_graph_client_service_principals)
 
     # RBAC related
     with self.command_group('ad sp', exception_handler=graph_err_handler) as g:
@@ -108,14 +108,14 @@ def load_command_table(self, _):
 
     with self.command_group('ad user', role_users_sdk, exception_handler=graph_err_handler) as g:
         g.command('delete', 'delete')
-        g.command('show', 'get')
+        g.show_command('show', 'get')
         g.custom_command('list', 'list_users', client_factory=get_graph_client_users)
         g.custom_command('create', 'create_user', client_factory=get_graph_client_users, doc_string_source='azure.graphrbac.models#UserCreateParameters')
 
     with self.command_group('ad group', role_group_sdk, exception_handler=graph_err_handler) as g:
         g.custom_command('create', 'create_group', client_factory=get_graph_client_groups)
         g.command('delete', 'delete')
-        g.command('show', 'get')
+        g.show_command('show', 'get')
         g.command('get-member-groups', 'get_member_groups')
         g.custom_command('list', 'list_groups', client_factory=get_graph_client_groups)