+++ src/azure-cli-core/HISTORY.rst
@@ -6,6 +6,7 @@ Release History
 2.0.42
 ++++++
 * login: support browser based login in WSL bash window
+* Adds `--force-string` flag to all generic update commands.
 
 2.0.41
 ++++++
@@ -34,7 +35,7 @@ Release History
 
 2.0.35
 ++++++
-* Minor fixes
+* Added method of registering `show` commands to fail with exit code of 3.
 
 2.0.34
 ++++++