+++ src/command_modules/azure-cli-lab/HISTORY.rst
@@ -5,7 +5,7 @@ Release History
 
 0.1.0
 +++++
-* Minor fixes
+* BREAKING CHANGE: 'show' commands log error message and fail with exit code of 3 upon a missing resource.
 
 0.0.23
 ++++++