+++ src/command_modules/azure-cli-resource/HISTORY.rst
@@ -3,15 +3,17 @@
 Release History
 ===============
 
-2.0.34
+2.1.1
 ++++++
 * Support OnError group deployments.
 
-2.0.33
-++++++
+2.1.0
++++++
+* BREAKING CHANGE: 'show' commands log error message and fail with exit code of 3 upon a missing resource.
 * `group deployment delete`: Add `--no-wait` support.
 * `deployment delete`: Add `--no-wait` support.
 * Added `deployment wait` command.
+* Fix issue where the subscription-level `az deployment` commands erroneously appeared for profile 2017-03-09-profile.
 
 2.0.32
 ++++++